+++ plugin_ide.ui/src/melnorme/lang/ide/ui/LangUIPlugin_Actual.java
@@ -18,6 +18,8 @@ public final class LangUIPlugin_Actual {
 	
 	public static final String PLUGIN_ID = DeeUIPlugin.PLUGIN_ID;
 	
+	public static final String ROOT_PREF_PAGE_ID = PLUGIN_ID + ".PreferencePages.Root";
+	
 	public static final String RULER_CONTEXT = "#DeeRulerContext";
 	public static final String EDITOR_CONTEXT = "#DeeEditorContext";