+++ plugin_ide.ui/src/mmrnmhrm/ui/preferences/pages/DeeFoldingPreferencePage.java
@@ -13,7 +13,7 @@ import org.eclipse.jface.preference.PreferencePage;
 
 public class DeeFoldingPreferencePage extends AbstractConfigurationBlockPreferencePage {
 	
-	public final static String PAGE_ID = DeeUIPlugin.PLUGIN_ID + ".preferences.editor.folding";
+	public final static String PAGE_ID = DeeUIPlugin.PLUGIN_ID + ".PreferencePages.Editor.Folding";
 	
 	@Override
 	protected void setDescription() {